Factor out parts of NavigationManager.ts that will be reused
